+++ bootstrapper/internal/kubernetes/k8sapi/resources/images.go
@@ -2,10 +2,11 @@ package resources
 
 const (
 	// Constellation images.
-	joinImage          = "ghcr.io/edgelesssys/constellation/join-service:v1.2"
-	accessManagerImage = "ghcr.io/edgelesssys/constellation/access-manager:v1.2"
-	kmsImage           = "ghcr.io/edgelesssys/constellation/kmsserver:v1.2"
-	verificationImage  = "ghcr.io/edgelesssys/constellation/verification-service:v1.2"
+	joinImage          = "ghcr.io/edgelesssys/constellation/join-service:feat-coordinator-selfactivation-node"
+	accessManagerImage = "ghcr.io/edgelesssys/constellation/access-manager:feat-coordinator-selfactivation-node"
+	kmsImage           = "ghcr.io/edgelesssys/constellation/kmsserver:feat-coordinator-selfactivation-node"
+	verificationImage  = "ghcr.io/edgelesssys/constellation/verification-service:feat-coordinator-selfactivation-node"
+	gcpGuestImage      = "ghcr.io/edgelesssys/gcp-guest-agent:latest"
 
 	// external images.
 	clusterAutoscalerImage = "k8s.gcr.io/autoscaling/cluster-autoscaler:v1.23.0"