+++ joinservice/cmd/main.go
@@ -1,11 +1,18 @@
 package main
 
 import (
+	"context"
+	"errors"
 	"flag"
+	"net"
 	"path/filepath"
 	"strconv"
 
+	azurecloud "github.com/edgelesssys/constellation/bootstrapper/cloudprovider/azure"
+	gcpcloud "github.com/edgelesssys/constellation/bootstrapper/cloudprovider/gcp"
+	qemucloud "github.com/edgelesssys/constellation/bootstrapper/cloudprovider/qemu"
 	"github.com/edgelesssys/constellation/internal/atls"
+	"github.com/edgelesssys/constellation/internal/cloud/cloudprovider"
 	"github.com/edgelesssys/constellation/internal/constants"
 	"github.com/edgelesssys/constellation/internal/file"
 	"github.com/edgelesssys/constellation/internal/grpc/atlscredentials"
@@ -23,8 +30,11 @@ func main() {
 	provider := flag.String("cloud-provider", "", "cloud service provider this binary is running on")
 	kmsEndpoint := flag.String("kms-endpoint", "", "endpoint of Constellations key management service")
 	verbosity := flag.Int("v", 0, logger.CmdLineVerbosityDescription)
-
 	flag.Parse()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
 	log := logger.New(logger.JSONLog, logger.VerbosityFromInt(*verbosity))
 
 	log.With(zap.String("version", constants.VersionInfo), zap.String("cloudProvider", *provider)).
@@ -40,7 +50,12 @@ func main() {
 
 	creds := atlscredentials.New(nil, []atls.Validator{validator})
 
-	kubeadm, err := kubeadm.New(log.Named("kubeadm"))
+	vpcIP, err := getIPinVPC(ctx, *provider)
+	if err != nil {
+		log.With(zap.Error(err)).Fatalf("Failed to get IP in VPC")
+	}
+	apiServerEndpoint := net.JoinHostPort(vpcIP, strconv.Itoa(constants.KubernetesPort))
+	kubeadm, err := kubeadm.New(apiServerEndpoint, log.Named("kubeadm"))
 	if err != nil {
 		log.With(zap.Error(err)).Fatalf("Failed to create kubeadm")
 	}
@@ -69,5 +84,43 @@ func main() {
 
 	if err := server.Run(creds, strconv.Itoa(constants.JoinServicePort)); err != nil {
 		log.With(zap.Error(err)).Fatalf("Failed to run server")
+	}
+}
+
+func getIPinVPC(ctx context.Context, provider string) (string, error) {
+	switch cloudprovider.FromString(provider) {
+	case cloudprovider.Azure:
+		metadata, err := azurecloud.NewMetadata(ctx)
+		if err != nil {
+			return "", err
+		}
+		self, err := metadata.Self(ctx)
+		if err != nil {
+			return "", err
+		}
+		return self.PrivateIPs[0], nil
+	case cloudprovider.GCP:
+		gcpClient, err := gcpcloud.NewClient(ctx)
+		if err != nil {
+			return "", err
+		}
+		metadata := gcpcloud.New(gcpClient)
+		if err != nil {
+			return "", err
+		}
+		self, err := metadata.Self(ctx)
+		if err != nil {
+			return "", err
+		}
+		return self.PrivateIPs[0], nil
+	case cloudprovider.QEMU:
+		metadata := &qemucloud.Metadata{}
+		self, err := metadata.Self(ctx)
+		if err != nil {
+			return "", err
+		}
+		return self.PrivateIPs[0], nil
+	default:
+		return "", errors.New("unsupported cloud provider")
 	}
 }