+++ CHANGELOG.md
@@ -22,9 +22,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Nodes add themselves to the cluster after `constellation init` is done
+Removed
+
 ### Deprecated
 
 ### Removed
+
+- User facing WireGuard VPN
 
 ### Fixed